+++ src/Terrabuild/Program.fs
@@ -813,6 +813,7 @@ let processCommandLine (parser: ArgumentParser<TerrabuildArgs>) (result: ParseRe
         if clearArgs.Contains(ClearArgs.Cache) || clearArgs.Contains(ClearArgs.All) then Cache.clearCache()
         if clearArgs.Contains(ClearArgs.Home) || clearArgs.Contains(ClearArgs.All) then Cache.clearHomeCache()
         if clearArgs.Contains(ClearArgs.Temporary) || clearArgs.Contains(ClearArgs.All) then Cache.clearTemp()
+        if clearArgs.Contains(ClearArgs.All) then TargetLock.clear()
         0
 
     let prune (pruneArgs: ParseResults<PruneArgs>) =